DiscordBot/pyproject.toml
[tool.poetry]
name = "discordbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
protobuf = "^4.25.1"
types-protobuf = "^4.24.0.4"
discord-py = "^2.3.2"
betterproto = { extras = [
  "compiler",
], version = "^2.0.0b6", allow-prereleases = true }
pytz = "^2023.3.post1"
schedule = "^1.2.1"


[tool.poetry.group.dev.dependencies]
types-protobuf = "^4.24.0.4"
grpc-stubs = "^1.53.0.3"
pyright = "^1.1.337"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
